requirements.txt
absl-py==2.1.0
filelock==3.17.0
fsspec==2025.2.0
grpcio==1.70.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
